--- test/__mocks__/weather.js
@@ -1,52 +1,0 @@
-const weatherData = {
-  'London': {
-    name: 'London',
-    main: {
-      temp: 275
-    },
-    coord: {
-      lat: 51.5085,
-      lon: -0.1257
-    }
-  }
-};
-
-const cities = {
-  'London': [
-    {
-      "name": "London",
-      "local_names": {
-        "ar": "لندن",
-        "en": "London",
-        "ru": "Лондон",
-      },
-      "lat": 51.5085,
-      "lon": -0.1257,
-      "country": "GB"
-    }
-  ]
-};
-
-const weather = {
-  get(lat, lon) {
-    return new Promise((resolve, reject) => {
-      lat > 0
-        ? resolve(weatherData)
-        : reject(new Error("Weather was not found"));
-    });
-  },
-
-  findCityCoords(cityName) {
-    return new Promise((resolve, reject) => {
-      cities[cityName]
-        ? resolve(cities[cityName][0])
-        : reject(new Error("City was not found"));
-    });
-  },
-
-  toCelsius(fahrenheit) {
-    return parseFloat(fahrenheit) - 273.15;
-  },
-};
-
-export default weather;
