+++ test/page.test.js
@@ -1,5 +1,4 @@
 import Page from "../src/page";
-import weather from "./__mocks__/weather";
 
 jest.mock('../src/weather');
 
@@ -72,17 +71,10 @@ describe('Page', () => {
     expect(buttons[2].innerHTML).toBe(history[2]);
   })
 
-  it('show city weather', () => {
-    weather.findCityCoords('London')
-      .then((cityCoords) => weather.get(cityCoords.lat, cityCoords.lon))
-      .then((weatherData) => {
-        console.log(weatherData);
-      })
-      .catch((err) => console.error(err));
-
-    //page.showCityWeather('London');
-    //expect(page.t.innerHTML).toBe('+2°C');
-    //expect(page.city.innerHTML).toBe('London');
+  it('show city weather', async () => {
+    await page.showCityWeather('London');
+    expect(page.t.innerHTML).toBe('+2°C');
+    expect(page.city.innerHTML).toBe('London');
   })
 
   it('locks page', () => {
